Fix static analysis warnings (#927)
* Fix the warnings

* Add comment

* Remove redundant function call

* Remove unused slice

+++ zk/stages/stage_sequence_execute_utils.go
@@ -435,7 +435,7 @@ func (bdc *BlockDataChecker) AddBlockStartData() bool {
 
 func (bdc *BlockDataChecker) AddTransactionData(txL2Data []byte) bool {
 	encodedLen := uint64(len(txL2Data))
-	if bdc.counter+uint64(encodedLen) > bdc.limit {
+	if bdc.counter+encodedLen > bdc.limit {
 		return true
 	}
